+++ src/problems/meta/coding/puzzles/warmup/sum_abc/__main__.py
@@ -1,20 +1,8 @@
 from hamcrest import assert_that, equal_to
 
-"""
-Given three integers A, B, and C, determine their sum.
-
-Your task is to implement the function getSum(A, B, C)
-which returns the sum A+B+C.
-"""
-
-
-def calculate_sum(a: int, b: int, c: int) -> int:
-    return a + b + c
-
+from src.problems.meta.coding.puzzles.warmup.sum_abc.v1 import getSum
 
 # pylint: disable=invalid-name
-def getSum(A: int, B: int, C: int) -> int:
-    return calculate_sum(A, B, C)
 
 
 def test(A: int, B: int, C: int, expected: int) -> None:
