+++ src/problems/meta/coding/puzzles/warmup/sum_abc/v1.py
@@ -1,0 +1,16 @@
+"""
+Given three integers A, B, and C, determine their sum.
+
+Your task is to implement the function getSum(A, B, C)
+which returns the sum A+B+C.
+"""
+
+
+def calculate_sum(a: int, b: int, c: int) -> int:
+    return a + b + c
+
+
+# pylint: disable=invalid-name
+def getSum(A: int, B: int, C: int) -> int:
+    return calculate_sum(A, B, C)
+# pylint: enable=invalid-name
